// src/app/models/invoice/invoice-adustment-data-view.ts
import { InvoiceAdjustmentListView } from './invoice-adjustment-list-view';

/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * InvoiceAdjustmentDataView
 */
export interface InvoiceAdjustmentDataView { 
    /**
     * Gets or Sets
     */
    invoiceId?: string;
    /**
     * Gets or Sets
     */
    invoiceAdjustmentListView?: Array<InvoiceAdjustmentListView>;
}