// src/app/models/invoice/payment-view-model.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * PaymentViewModel
 */
export interface PaymentViewModel { 
    /**
     * Gets or Sets
     */
    studentId?: string;
    /**
     * Gets or Sets
     */
    paymentDate?: Date;
    /**
     * Gets or Sets InvoiceIds
     */
    invoiceIds?: Array<string>;
    /**
     * Gets or Sets
     */
    feeAccountId?: number;
    /**
     * Gets or Sets
     */
    typeCode?: string;
    /**
     * Gets or Sets
     */
    cardTypeId?: number;
    /**
     * Gets or Sets
     */
    cardNumber?: string;
    /**
     * Gets or Sets
     */
    amount?: number;
    /**
     * Gets or Sets
     */
    transactionNumber?: string;
    /**
     * Gets or Sets
     */
    authorizationCode?: string;
    /**
     * Gets or Sets
     */
    chequeNumber?: string;
    /**
     * Gets or Sets
     */
    chequeDate?: Date;
    /**
     * Gets or Sets
     */
    chequeIssueBankName?: string;
    /**
     * Gets or Sets
     */
    checkIssueBankIFSC?: string;
    /**
     * Gets or Sets
     */
    chequeIssueBankBranch?: string;
        /**
     * Gets or Sets InvoiceIds
     */
    schoolAcademicSessionId: number;
}